+++ src/components/card-offer-list/card-offer-list.tsx
@@ -1,5 +1,11 @@
-import type {ListOfferProps} from '../../types.ts';
+import type {OfferPreview} from '../../types.ts';
 import CardOffer from '../card-offer/card-offer.tsx';
+
+type ListOfferProps = {
+  listOffer: OfferPreview[];
+  variantCard : 'cities'|'favorite';
+  mouseMove: (id:string|null) => void;
+}
 
 export default function ListOffer({listOffer,variantCard, mouseMove} : ListOfferProps) : JSX.Element {
   return (
