+++ src/components/card-offer/card-offer.tsx
@@ -1,5 +1,11 @@
-import type {OfferProps} from '../../types.ts';
+import type {OfferPreview} from '../../types.ts';
 import {Link} from 'react-router-dom';
+
+type OfferProps = {
+  offer: OfferPreview;
+  variantCard : 'cities'|'favorite';
+  mouseMove: (id:string|null) => void;
+}
 
 export default function CardOffer ({offer,variantCard,mouseMove} : OfferProps) : JSX.Element {
